Rename dirs as per PR comment
